port: phase 5 -- persistence via System.Text.Json + SaveStorage abstraction
Replace fastJSON + IsolatedStorageFile with System.Text.Json behind an
ISaveStorage abstraction (FileSaveStorage -> %AppData%/MightyFights/SaveData.json,
injectable via DataStore.cSaveStorage for the future web head). [JsonIgnore] the
two derived lookup dictionaries (Steward.hCompaniesBy*) and Company.caTemplates so
HydrateCompanyRefLists rebuilds refs without a duplicate-key crash. Verified by an
isolated round-trip harness (14/14 checks: byte-identical JSON, reference identity,
enum/string dicts).

Also fixes two latent XNA->MonoGame errors that a broken fastJSON using had been
masking (MessageBox in TitleScreen, Color overload ambiguity in CompnayMenu). The
full solution now compiles with 0 errors for the first time in the port.

Ignore .vs/ (VS working cache, absent from the original XNA-era .gitignore).

+++ src/MightyFights.Core/Scenes/InGame/Camp/Objects/Menus/Company Menu/CompnayMenu.cs
@@ -294,7 +294,9 @@ namespace MightyFights_Prototype
 					@"Sprite Data\Troopers\Halberd\Textures\fazure", new Dictionary<string, bool>());
 				_cGuy.cTexRef = cContent.Load<Texture2D>(@"Sprite Data\Troopers\Halberd\Front Facing\front");
 				_cGuy.fZRange = .3f;
-				_cGuy.tColor = new Color(_cGuy.tColor.R, _cGuy.tColor.G, _cGuy.tColor.B, 50);
+				// PORT (Phase 5): alpha cast to byte to disambiguate MonoGame's Color(int,int,int,int)
+				// vs Color(byte,byte,byte,byte) overloads (the int literal 50 matches both).
+				_cGuy.tColor = new Color(_cGuy.tColor.R, _cGuy.tColor.G, _cGuy.tColor.B, (byte)50);
 				// the scale throws this off quite a bit
 				_cGuy.tPos = new Vector2(tPos.X - 30, tPos.Y - 5);
